Separate Document Ready from Document State
