+++ extensions/amp-access/0.1/amp-login-done.js
@@ -22,7 +22,7 @@
 import '../../../third_party/babel/custom-babel-helpers';
 import '../../../src/polyfills';
 import {LoginDoneDialog} from './amp-login-done-dialog';
-import {onDocumentReady} from '../../../src/document-state';
+import {onDocumentReady} from '../../../src/document-ready';
 
 onDocumentReady(document, () => {
   new LoginDoneDialog(window).start();
